requirements.txt
PyMySQL==0.10.1
python-dotenv==0.15.0
